Remove old ImportFileJob code
