+++ app/jobs/import_from_ftp_job.rb
@@ -7,10 +7,10 @@ class ImportFromFtpJob < ApplicationJob
       source ETL::Sources::Ftp, ftp_credential_id: ftp_credential_id
 
       # Skip unhandled rows.
-      transform ETL::Transformations::SkipUnwantedRowsFtp
+      transform ETL::Transformations::SkipUnwantedRows
 
       # Create a job for each row, on a queue for that type.
-      destination ETL::Destinations::ImportQueuesFtp
+      destination ETL::Destinations::ImportQueues
     end
 
     Kiba.run(job)
