+++ app/services/etl/transformations/skip_unwanted_rows.rb
@@ -1,6 +1,7 @@
 class ETL::Transformations::SkipUnwantedRows
-  def process(row)
-    row unless unwanted.include?(row[0])
+  # args are { row: [], filename: "" }
+  def process(args)
+    args unless unwanted.include?(args[:row][0])
   end
 
   private
