+++ spec/jobs/fake_data_job/fake_data_spec.rb
@@ -1,4 +1,4 @@
-describe ImportFileJob, type: :job do
+describe ImportFromFtpJob, type: :job do
   let!(:admin) { create(:confirmed_user) }
 
   let(:classes) { ['AbsenceRecord'] }
