+++ spec/jobs/import_from_ftp_job/costing_record_spec.rb
@@ -1,42 +1,42 @@
-describe ImportFileJob, type: :job do
+describe ImportFromFtpJob, type: :job do
   let!(:admin) { create(:confirmed_user) }
   let(:ftp_credential) { create(:ftp_credential, path: path) }
   let(:import_job) { ImportFromFtpJob.perform_later(ftp_credential_id: ftp_credential.id) }
 
   describe 'add' do
-    let(:path) { 'good_imports/absence_record/add' }
+    let(:path) { 'good_imports/costing_record/add' }
 
     it 'queues the job' do
       expect { import_job }
         .to change(ActiveJob::Base.queue_adapter.enqueued_jobs, :size).by(1)
     end
 
-    it 'creates a new AbsenceRecord' do
+    it 'creates a new CostingRecord' do
       perform_enqueued_jobs { import_job }
 
-      expect(AbsenceRecord.count).to eq(1)
-      pr = AbsenceRecord.first
+      expect(CostingRecord.count).to eq(1)
+      pr = CostingRecord.first
 
-      # Expect values in the database to match input from add_absence_record_20201015_00001157.DAT.
-      Expectations::AbsenceRecord.added.each do |key, value|
+      # Expect values in the database to match input from add_costing_record_20201015_00001157.DAT.
+      Expectations::CostingRecord.added.each do |key, value|
         expect(pr.send(key)).to eq(value)
       end
     end
   end
 
   describe 'update' do
-    let(:path) { 'good_imports/absence_record/update' }
+    let(:path) { 'good_imports/costing_record/update' }
 
     it 'queues the job' do
       expect { import_job }
         .to change(ActiveJob::Base.queue_adapter.enqueued_jobs, :size).by(1)
     end
 
-    context 'with an existing AbsenceRecord' do
+    context 'with an existing CostingRecord' do
       before do
-        build(:absence_record) do |r|
+        build(:costing_record) do |r|
           r.assign_attributes(
-            Expectations::AbsenceRecord.added.merge(
+            Expectations::CostingRecord.added.merge(
               'created_at' => Time.current - 1.week,
               'updated_at' => Time.current - 1.week
             )
@@ -48,14 +48,14 @@ describe ImportFileJob, type: :job do
       it 'updates it' do
         perform_enqueued_jobs { import_job }
 
-        expect(AbsenceRecord.count).to eq(1)
-        pr = AbsenceRecord.first
+        expect(CostingRecord.count).to eq(1)
+        pr = CostingRecord.first
 
         expect(pr.created_at).to be_within(2.seconds).of(Time.current - 1.week)
         expect(pr.updated_at).to be_within(2.seconds).of(Time.current)
 
-        # Expect values in the database to match input from update_absence_record_20201015_00001157.DAT.
-        Expectations::AbsenceRecord.updated.each do |key, value|
+        # Expect values in the database to match input from update_costing_record_20201015_00001157.DAT.
+        Expectations::CostingRecord.updated.each do |key, value|
           expect(pr.send(key)).to eq(value)
         end
       end
@@ -63,19 +63,19 @@ describe ImportFileJob, type: :job do
       it 'versions it' do
         perform_enqueued_jobs { import_job }
 
-        expect(AbsenceRecord.first.versions.last.whodunnit_type).to eq('Import')
-        expect(AbsenceRecord.first.versions.last.whodunnit).to include('update')
+        expect(CostingRecord.first.versions.last.whodunnit_type).to eq('Import')
+        expect(CostingRecord.first.versions.last.whodunnit).to include('update')
       end
     end
   end
 
   describe 'delete' do
-    let(:path) { 'good_imports/absence_record/delete' }
+    let(:path) { 'good_imports/costing_record/delete' }
 
     it 'deletes it' do
       perform_enqueued_jobs { import_job }
 
-      expect(AbsenceRecord.count).to eq(0)
+      expect(CostingRecord.count).to eq(0)
     end
   end
 end
