+++ spec/jobs/import_from_ftp_job/iso_8859_1_encoding_spec.rb
