+++ spec/jobs/import_from_ftp_job/mixed_import_spec.rb
