+++ spec/jobs/import_from_ftp_job/no_admins_spec.rb
@@ -1,4 +1,4 @@
-describe ImportFileJob, type: :job do
+describe ImportFromFtpJob, type: :job do
   let(:unconfirmed_user) { create(:user) { |u| u.update_column(:confirmed_at, nil) } }
   let(:ftp_credential) { create(:ftp_credential, path: path, created_by_id: unconfirmed_user.id) }
   let(:path) { 'good_imports/mixed_import' }
