+++ spec/jobs/import_from_ftp_job/qualification_record_spec.rb
@@ -1,42 +1,42 @@
-describe ImportFileJob, type: :job do
+describe ImportFromFtpJob, type: :job do
   let!(:admin) { create(:confirmed_user) }
   let(:ftp_credential) { create(:ftp_credential, path: path) }
   let(:import_job) { ImportFromFtpJob.perform_later(ftp_credential_id: ftp_credential.id) }
 
   describe 'add' do
-    let(:path) { 'good_imports/person_eit_record/add' }
+    let(:path) { 'good_imports/qualification_record/add' }
 
     it 'queues the job' do
       expect { import_job }
         .to change(ActiveJob::Base.queue_adapter.enqueued_jobs, :size).by(1)
     end
 
-    it 'creates a new PersonEitRecord' do
+    it 'creates a new QualificationRecord' do
       perform_enqueued_jobs { import_job }
 
-      expect(PersonEitRecord.count).to eq(1)
-      pr = PersonEitRecord.first
+      expect(QualificationRecord.count).to eq(1)
+      pr = QualificationRecord.first
 
-      # Expect values in the database to match input from add_person_eit_record_20201015_00001157.DAT.
-      Expectations::PersonEitRecord.added.each do |key, value|
+      # Expect values in the database to match input from add_qualification_record_20201015_00001157.DAT.
+      Expectations::QualificationRecord.added.each do |key, value|
         expect(pr.send(key)).to eq(value)
       end
     end
   end
 
   describe 'update' do
-    let(:path) { 'good_imports/person_eit_record/update' }
+    let(:path) { 'good_imports/qualification_record/update' }
 
     it 'queues the job' do
       expect { import_job }
         .to change(ActiveJob::Base.queue_adapter.enqueued_jobs, :size).by(1)
     end
 
-    context 'with an existing PersonEitRecord' do
+    context 'with an existing QualificationRecord' do
       before do
-        build(:person_eit_record) do |r|
+        build(:qualification_record) do |r|
           r.assign_attributes(
-            Expectations::PersonEitRecord.added.merge(
+            Expectations::QualificationRecord.added.merge(
               'created_at' => Time.current - 1.week,
               'updated_at' => Time.current - 1.week
             )
@@ -48,14 +48,14 @@ describe ImportFileJob, type: :job do
       it 'updates it' do
         perform_enqueued_jobs { import_job }
 
-        expect(PersonEitRecord.count).to eq(1)
-        pr = PersonEitRecord.first
+        expect(QualificationRecord.count).to eq(1)
+        pr = QualificationRecord.first
 
         expect(pr.created_at).to be_within(2.seconds).of(Time.current - 1.week)
         expect(pr.updated_at).to be_within(2.seconds).of(Time.current)
 
-        # Expect values in the database to match input from update_person_eit_record_20201015_00001157.DAT.
-        Expectations::PersonEitRecord.updated.each do |key, value|
+        # Expect values in the database to match input from update_qualification_record_20201015_00001157.DAT.
+        Expectations::QualificationRecord.updated.each do |key, value|
           expect(pr.send(key)).to eq(value)
         end
       end
@@ -63,19 +63,19 @@ describe ImportFileJob, type: :job do
       it 'versions it' do
         perform_enqueued_jobs { import_job }
 
-        expect(PersonEitRecord.first.versions.last.whodunnit_type).to eq('Import')
-        expect(PersonEitRecord.first.versions.last.whodunnit).to include('update')
+        expect(QualificationRecord.first.versions.last.whodunnit_type).to eq('Import')
+        expect(QualificationRecord.first.versions.last.whodunnit).to include('update')
       end
     end
   end
 
   describe 'delete' do
-    let(:path) { 'good_imports/person_eit_record/delete' }
+    let(:path) { 'good_imports/qualification_record/delete' }
 
     it 'deletes it' do
       perform_enqueued_jobs { import_job }
 
-      expect(PersonEitRecord.count).to eq(0)
+      expect(QualificationRecord.count).to eq(0)
     end
   end
 end
